done recommendation doctor view
