+++ lib/core/utils/app_router.dart
@@ -1,6 +1,6 @@
 import 'package:doc_doc_app/core/utils/routes.dart';
 import 'package:doc_doc_app/features/home/presentation/views/home_view.dart';
-import 'package:doc_doc_app/features/home/presentation/views/recommendation_view.dart';
+import 'package:doc_doc_app/features/recommendation/presentation/views/recommendation_view.dart';
 import 'package:doc_doc_app/features/home/presentation/views/speciality_view.dart';
 import 'package:doc_doc_app/features/login/presentation/views/login_view.dart';
 import 'package:doc_doc_app/features/onboarding/presentation/views/onboarding_view.dart';
