+++ lib/features/home/presentation/views/widgets/custom_doctor_rating.dart
@@ -1,7 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-
 import '../../../../../core/utils/color.dart';
 import '../../../../../core/utils/font_weight_helper.dart';
 import '../../../../../core/utils/spacer.dart';
