+++ lib/features/recommendation/presentation/views/recommendation_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'widgets/recommendation_view_body.dart';
+import 'widget/recommendation_view_body.dart';
 
 class RecommendationView extends StatelessWidget {
   const RecommendationView({super.key});
